requirements.txt
icalendar==3.10
python-dateutil==2.5.3
pytodoist==2.0.4
pytz==2016.6.1
PyYAML==3.12
requests==2.11.1
six==1.10.0
